fix compilation errors on 2017b

+++ external/toolbox_fast_marching/mex/perform_front_propagation_3d_mex.cpp
@@ -83,7 +83,7 @@ void mexFunction(	int nlhs, mxArray *plhs[],
 		values = NULL;
 		
 	// first ouput : distance
-	int dims[3] = {n,p,q};
+	size_t dims[3] = {n,p,q};
 	plhs[0] = mxCreateNumericArray(3, dims, mxDOUBLE_CLASS, mxREAL );
 	D = mxGetPr(plhs[0]);
 	// second output : state
